feature: buffer printing and more
printing now uses a buffer
min max now use long long instead of int
some extra files for printing functions and removed empty files
updated makefile and xcode project

+++ ft_printf/Source/libft/ft_min.c
@@ -10,7 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_min(int x, int y)
+long long	ft_min(long long x, long long y)
 {
 	if (x <= y)
 		return (x);
